Fix prototype of poll
from poll(struct pollfd *, nfds_t, long) to
poll(struct pollfd *, nfds_t, int)

See 11952073af22568bba0b661f7a9d4402c443a888
and ebfbb367fbba29e825e1cf2beb88800e05e9778f

+++ ndk/sources/android/libportable/arch-mips/poll.c
@@ -106,7 +106,7 @@ static inline short change_mips_events(short mips_events)
 
 extern int poll(struct pollfd *, nfds_t, int);
 
-int WRAP(poll)(struct pollfd *fds, nfds_t nfds, long timeout)
+int WRAP(poll)(struct pollfd *fds, nfds_t nfds, int timeout)
 {
   nfds_t i;
   int ret;
